refac: moved timeseries controller to readings.controllers.ts
